video_to_gif/requirements.txt
moviepy==1.0.3
click==8.1.3
Pillow==9.5.0
tqdm==4.65.0
pygifsicle==1.0.7
numpy==1.24.3
